홈베이스 예약 멤버 명단 수정하기

+++ src/main/java/msg/team1/Hi/domain/home_base/presentation/dto/request/ReserveHomeBaseRequest.java
@@ -20,5 +20,5 @@ public class ReserveHomeBaseRequest {
     @NotNull
     private Integer period;
     @NotEmpty
-    private List<Integer> members;
+    private List<Long> members;
 }
